+++ src/main/java/fossilsarcheology/client/render/entity/layer/LayerAnuStatueGlow.java
@@ -10,18 +10,15 @@ import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
 
 import java.util.Random;
 
-public class LayerAnuStatueGlow implements LayerRenderer<EntityAnuStatue>
-{
-    public void doRenderLayer(EntityAnuStatue anu, float limbSwing, float limbSwingAmount, float partialTicks, float ageInTicks, float netHeadYaw, float headPitch, float scale)
-    {
+public class LayerAnuStatueGlow implements LayerRenderer<EntityAnuStatue> {
+    public void doRenderLayer(EntityAnuStatue anu, float limbSwing, float limbSwingAmount, float partialTicks, float ageInTicks, float netHeadYaw, float headPitch, float scale) {
         Tessellator tessellator = Tessellator.getInstance();
         BufferBuilder bufferbuilder = tessellator.getBuffer();
         RenderHelper.disableStandardItemLighting();
-        float f = ((float)anu.ticksExisted + partialTicks) / 200.0F;
+        float f = ((float) anu.ticksExisted + partialTicks) / 200.0F;
         float f1 = 0.0F;
 
-        if (f > 0.8F)
-        {
+        if (f > 0.8F) {
             f1 = (f - 0.8F) / 0.2F;
         }
 
@@ -34,8 +31,7 @@ public class LayerAnuStatueGlow implements LayerRenderer<EntityAnuStatue>
         GlStateManager.enableCull();
         GlStateManager.depthMask(false);
         GlStateManager.pushMatrix();
-        for (int i = 0; (float)i < (f + f * f) / 2.0F * 60.0F; ++i)
-        {
+        for (int i = 0; (float) i < (f + f * f) / 2.0F * 60.0F; ++i) {
             GlStateManager.rotate(random.nextFloat() * 360.0F, 1.0F, 0.0F, 0.0F);
             GlStateManager.rotate(random.nextFloat() * 360.0F, 0.0F, 1.0F, 0.0F);
             GlStateManager.rotate(random.nextFloat() * 360.0F, 0.0F, 0.0F, 1.0F);
@@ -45,11 +41,11 @@ public class LayerAnuStatueGlow implements LayerRenderer<EntityAnuStatue>
             float f2 = random.nextFloat() * 20.0F + 5.0F + f1 * 10.0F;
             float f3 = random.nextFloat() * 2.0F + 1.0F + f1 * 2.0F;
             bufferbuilder.begin(6, DefaultVertexFormats.POSITION_COLOR);
-            bufferbuilder.pos(0.0D, 0.0D, 0.0D).color(66, 0, 176, (int)(255.0F * (1.0F - f1))).endVertex();
-            bufferbuilder.pos(-0.866D * (double)f3, (double)f2, (double)(-0.5F * f3)).color(255, 0, 0, 0).endVertex();
-            bufferbuilder.pos(0.866D * (double)f3, (double)f2, (double)(-0.5F * f3)).color(255, 0, 0, 0).endVertex();
-            bufferbuilder.pos(0.0D, (double)f2, (double)(1.0F * f3)).color(255, 0, 0, 0).endVertex();
-            bufferbuilder.pos(-0.866D * (double)f3, (double)f2, (double)(-0.5F * f3)).color(255, 0, 0, 0).endVertex();
+            bufferbuilder.pos(0.0D, 0.0D, 0.0D).color(66, 0, 176, (int) (255.0F * (1.0F - f1))).endVertex();
+            bufferbuilder.pos(-0.866D * (double) f3, f2, -0.5F * f3).color(255, 0, 0, 0).endVertex();
+            bufferbuilder.pos(0.866D * (double) f3, f2, -0.5F * f3).color(255, 0, 0, 0).endVertex();
+            bufferbuilder.pos(0.0D, f2, 1.0F * f3).color(255, 0, 0, 0).endVertex();
+            bufferbuilder.pos(-0.866D * (double) f3, f2, -0.5F * f3).color(255, 0, 0, 0).endVertex();
             tessellator.draw();
         }
 
@@ -64,8 +60,7 @@ public class LayerAnuStatueGlow implements LayerRenderer<EntityAnuStatue>
         RenderHelper.enableStandardItemLighting();
     }
 
-    public boolean shouldCombineTextures()
-    {
+    public boolean shouldCombineTextures() {
         return false;
     }
 }
