+++ src/main/java/fossilsarcheology/server/entity/ai/DinoAIFlockWander.java
@@ -118,6 +118,6 @@ public class DinoAIFlockWander<T extends EntityPrehistoric> extends EntityAIBase
         float f1 = MathHelper.sin(-yaw * 0.017453292F - (float) Math.PI);
         float f2 = -MathHelper.cos(-pitch * 0.017453292F);
         float f3 = MathHelper.sin(-pitch * 0.017453292F);
-        return new Vec3d((double) (f1 * f2), (double) f3, (double) (f * f2));
+        return new Vec3d(f1 * f2, f3, f * f2);
     }
 }
